Update main.c
The reason why I changed this code is because the code did not produce 32 "Hello School. The integers were correct but i had to dig a little bit to know I needed the initial C command of " #include stdio.h ". Thats it. But that polished a clean "hello School" 32 times in the making.